+++ src/markehme/factionsplus/config/sections/SubSection_Protection.java
@@ -10,6 +10,10 @@ public class SubSection_Protection {
 			realAlias_neverDotted = "LWC" )
 	public final SubSection_LWC	_lwc							= new SubSection_LWC();
 	
+	@Section(
+		realAlias_neverDotted = "PVP" )
+	public final SubSection_PVP	_pvp							= new SubSection_PVP();
+	
 	// FIXME: this interferes with LWC's option
 	@Option(autoComment={"unused"},
 			realAlias_inNonDottedFormat = "removeSignProtectionOnClaim" )