+++ src/markehme/factionsplus/listeners/CoreListener.java
@@ -21,6 +21,7 @@ import com.massivecraft.factions.event.FPlayerLeaveEvent;
 import com.massivecraft.factions.event.FactionDisbandEvent;
 
 public class CoreListener implements Listener{
+	
 	@EventHandler
 	public void onPlayerCommandPreprocess(PlayerCommandPreprocessEvent event) {
 		if(event.isCancelled()) {